requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
Werkzeug==2.3.7
google-auth==2.22.0
requests==2.31.0
openai>=1.50.0
Flask-Limiter==3.3.1
Flask-Caching==2.0.2
gunicorn==21.2.0
whitenoise==6.5.0
flasgger==0.9.7.1
pytesseract==0.3.10
Pillow>=10.3.0
pytest==7.4.0
pytest-flask==1.2.0
